# Speed up command line help