Speed up command line help by avoiding importing heavy modules and reinstate Google docstring processing after #1101 broke it.

| command | master[*] | this branch |
|-|:-:|:-:|
| `improver help` | 4.438s | 2.977s |
| `improver help wxcode` | 2.733s | 1.257s |

[*] after applying fix to `clizefy` from this PR.

Testing:
 - [x] Ran tests and they passed OK
 - [x] Added new tests for fixed bug