@benfitzpatrick 2020-01-28
![Codacy](https://app.codacy.com/assets/images/favicon.png) Here is an overview of what got changed by this pull request:


```diff

Issues
======
+ Solved 2
- Added 1
           

Complexity decreasing per file
==============================
+ improver/wxcode/utilities.py  -3
         
```


See the complete overview on [Codacy](https://app.codacy.com/manual/metoppv_tech/improver/pullRequest?prid=4888025&bid=16191239)