@TomekTrzeciak 2020-01-29
I was going to fix https://github.com/metoppv/improver/pull/1132#discussion_r371739764 while I'm at it, but unfortunately it's not clear to me what's the right definition in some cases. Therefore I'm going to leave it for someone with the right expertise to fix it.

To help you out, here's the list of broken queries:
```
bad query for high_resolution tree: heavy_sleet_continuous
bad query for high_resolution tree: heavy_sleet_shower
bad query for high_resolution tree: heavy_rain_or_snow_continuous
bad query for high_resolution tree: heavy_rain_or_snow_shower
bad query for high_resolution tree: light_sleet_continuous
bad query for high_resolution tree: light_rain_or_snow_continuous
bad query for high_resolution tree: light_sleet_shower
bad query for high_resolution tree: light_rain_or_snow_shower
bad query for high_resolution tree: sleet_in_vicinity
bad query for high_resolution tree: rain_or_snow_in_vicinity
bad query for global tree: heavy_sleet_continuous
bad query for global tree: heavy_sleet_shower
bad query for global tree: heavy_rain_or_snow_continuous
bad query for global tree: heavy_rain_or_snow_shower
bad query for global tree: light_sleet_continuous
bad query for global tree: light_rain_or_snow_continuous
bad query for global tree: light_sleet_shower
bad query for global tree: light_rain_or_snow_shower
```

The above list was generated by filtering the queries against the following check:
```python
def check_query(query):
    def check_value_consistency(values):
        type_set = set(map(type, values))
        if list in type_set:
            return (
                len(type_set) == 1 and
                len(set(map(len, values))) == 1 and
                all(map(check_value_consistency, zip(*values)))
            )
        return True

    mandatory_keys = [
        'diagnostic_fields',
        'diagnostic_conditions',
        'diagnostic_thresholds',
        'probability_thresholds',
        'diagnostic_gamma',
    ]
    optional_keys = ['diagnostic_gamma']
    values_to_match = [query[key]
                       for key in mandatory_keys + optional_keys
                       if key not in optional_keys or key in query]

    return check_value_consistency(values_to_match)
```

You could stick it somewhere in wxcode plugin (or tests) to ensure consistency in the future, but instead I would encourage you to redesign the data structures used for queries. Happy to discuss further if there's interest.